// src/shared/enums/perso-info-response.enum.ts
export enum InfoPersoReponse
{
  PASSWORD_NO_MATCH = 'PASSWORD_NO_MATCH',
  NEW_PASSWORD_NO_MATCH = 'NEW_PASSWORD_NO_MATCH', 
  PASSWORD_CHANGED_SUCCESS = 'PASSWORD_CHANGED_SUCCESS',
  MOBILE_NO_MATCH_FORMAT = 'MOBILE_NO_MATCH_FORMAT',
  MOBILE_CHANGED_SUCCESS = 'MOBILE_CHANGED_SUCCESS',
  DAMAGE_REPORT_SAVE = 'DAMAGE_REPORT_SAVE',
  DAMAGE_REPORT_FAIL = 'DAMAGE_REPORT_FAIL',
  PLATE_NUMBER_NO_MATCH = 'PLATE_NUMBER_NO_MATCH',
  CONTTRACT_NO_EXIST = 'CONTTRACT_NO_EXIST',


}